Switch timer to set OC0B on compare match mode
Before with clear on compare match, the min duty cycle was 1/256 and the
max 256/256. Now the min is 0/256 and the max is 255/256.

I want to be able to have true zero brightness for indications such as
binary output. Also, it may be useful to support zero brightness while
the driver is powered for future functionality such as a timer mode.