Remove special function brightness set on-time
Remove support for having the special function mode brightness set to
the current brightness automatically after a configured on-time. I do
not think this is actually that useful, since I find that most of the
time I prefer to just use a fixed, low brightness for special function
mode.

This was done to save program space. This commit gets the program size
down to 1020 bytes.